test: move scopes to files
